+++ src/components/Icon/index.tsx
@@ -1,2 +1,2 @@
 export { default } from "./Icon";
-export type { IconName } from "./Icon";
+export type { IconName, IconSize } from "./Icon";