+++ src/components/Svg/Svg.tsx
@@ -1,20 +1,17 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import color, { ColorProps } from "../../styles/utils/color";
-import display, { DisplayProps } from "../../styles/utils/display";
+import { box, BoxProps } from "../Box";
 import getSvgId, { SvgName } from "../SpriteSheet/getSvgId";
 
-const StyledSvg = styled.svg`
-  ${color}
-  ${display}
+const StyledSvg = styled.svg<BoxProps>`
+  ${box}
 `;
-type SvgProps = DisplayProps &
-  ColorProps & {
-    name: SvgName;
-    className?: string;
-    hideOnMobile?: boolean;
-  };
+type SvgProps = BoxProps & {
+  name: SvgName;
+  className?: string;
+  hideOnMobile?: boolean;
+};
 const Svg: FC<SvgProps> = (props) => {
   return (
     <StyledSvg